+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - New Error: EBUSYError #177
 
+### Changed
+- Dynamic p2pcommons URLs #178
+
 ## [0.6.1] - 2020-07-07
 ### Fixed
 - loading issue